Rename base_command to frame
