+++ msmart/base_device.py
@@ -2,8 +2,8 @@ import logging
 import time
 from typing import List, Optional
 
-from msmart.base_command import Frame
 from msmart.const import DeviceType
+from msmart.frame import Frame
 from msmart.lan import LAN, AuthenticationError, Key, ProtocolError, Token
 
 _LOGGER = logging.getLogger(__name__)
