+++ msmart/device/AC/command.py
@@ -8,8 +8,8 @@ from enum import IntEnum
 from typing import Callable, Collection, Mapping, Optional, Union
 
 import msmart.crc8 as crc8
-from msmart.base_command import Frame, InvalidFrameException
 from msmart.const import DeviceType, FrameType
+from msmart.frame import Frame, InvalidFrameException
 
 _LOGGER = logging.getLogger(__name__)
